1. joinRoom时获取对端appShowName/Desc； 2. 调整roomDuration为float类型；

+++ openvidu-server/src/main/java/io/openvidu/server/core/SessionPreset.java
@@ -77,7 +77,7 @@ public class SessionPreset {
                          String sharePowerInRoom,
                          String subject,
                          Integer roomCapacity,
-                         Integer roomDuration,
+                         Float roomDuration,
                          String useIdTypeInRoom,
                          String allowPartOperMic,
                          String allowPartOperShare) {
